requirements.txt
pandas
scikit-learn
xgboost
torch
torchvision
Pillow
